Move service tests to command tests
